% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_frame.R
\name{file_frame}
\alias{file_frame}
\title{Enhanced Working Directory View}
\usage{
file_frame(dir = ".")
}
\arguments{
\item{dir}{character, the directory in which to view files}
}
\value{
a data.frame of the directories and files names in presentation format
}
\description{
Enhanced Working Directory View
}
\examples{
file_frame()

}

