% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf.R
\name{lf}
\alias{lf}
\title{Enhanced Working Directory View}
\usage{
lf(dir = ".")
}
\arguments{
\item{dir}{character, the directory in which to view files}
}
\value{
a data.frame of the directories and files names in presentation format
}
\description{
Enhanced Working Directory View
}
\examples{
lf()

}

