% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df}
\alias{df}
\title{Describe Files}
\usage{
df(dir = ".", all.files = TRUE, ...)
}
\arguments{
\item{dir}{the directory containing the files to describe}

\item{all.files}{logical, include hidden files and folders}

\item{...}{additional arguments to pass to \code{list.files}}
}
\value{
A data.frame containing file information for all files in \code{dir}
}
\description{
Describe Files
}
\examples{
df()

}

